+++ app/src/main/java/com/example/demoproject/fragments/NotificationsFragment.kt
@@ -1,12 +1,16 @@
-package com.example.demoproject
+package com.example.demoproject.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.demoproject.R
 
-class SettingFragment : Fragment() {
+class NotificationsFragment : Fragment() {
+    // TODO: Rename and change types of parameters
+    private var param1: String? = null
+    private var param2: String? = null
 
 
     override fun onCreateView(
@@ -14,7 +18,7 @@ class SettingFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_setting, container, false)
+        return inflater.inflate(R.layout.fragment_notifications, container, false)
     }
 
 }
