+++ app/src/main/java/com/example/demoproject/fragments/SettingFragment.kt
@@ -1,15 +1,13 @@
-package com.example.demoproject
+package com.example.demoproject.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.demoproject.R
 
-class NotificationsFragment : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+class SettingFragment : Fragment() {
 
 
     override fun onCreateView(
@@ -17,7 +15,7 @@ class NotificationsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_notifications, container, false)
+        return inflater.inflate(R.layout.fragment_setting, container, false)
     }
 
 }
