+++ app/src/main/res/layout/activity_tab_layout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TabLayoutClass">
+    tools:context=".activities.TabLayoutClass">
 
 
     <androidx.viewpager.widget.ViewPager
