+++ app/src/main/res/layout/activity_training.xml
@@ -8,7 +8,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TrainingActivity">
+    tools:context=".activities.TrainingActivity">
 
     <TextView
         android:id="@+id/textView3"
