+++ app/src/main/res/layout/fragment_home.xml
@@ -10,7 +10,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#EFF2F9"
-    tools:context=".HomeFragment">
+    tools:context=".fragments.HomeFragment">
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/Rv"
         android:layout_width="347dp"
